pyproject.toml
[project]
name = "agents-hack-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.4",
    "langchain-google-genai>=3.0.1",
    "langchain-google-vertexai>=3.0.2",
    "python-dotenv>=1.2.1",
]
